Consolidate web and React Native call object loaders into one CallObjectLoader
